Convert float to size_t

+++ include/freddy/detail/manager.hpp
@@ -560,7 +560,7 @@ class manager
         gc();
         if (ut.load_factor() <= ut.max_load_factor() * (1.0f - config::dead_factor))
         {
-            ut.rehash(std::ceil(0.5f * ut.bucket_count()));
+            ut.rehash(static_cast<std::size_t>(std::ceil(0.5f * ut.bucket_count())));
         }
         else
         {  // too few nodes/edges of this UT were deleted
@@ -626,8 +626,7 @@ class manager
         // look ahead to determine how many swaps need to be performed
         auto const max_swaps_needed =
             std::accumulate(vl[x].nt.begin(), vl[x].nt.end(), 0, [swap_is_needed](auto const sum, auto const& v) {
-                // two branches must be made per swap
-                return swap_is_needed(v->br().hi, v->br().lo) ? sum + 2 : sum;
+                return swap_is_needed(v->br().hi, v->br().lo) ? sum + 2 : sum;  // two branches must be made per swap
             });
         if (vl[x].nt.size() + max_swaps_needed > vl[x].nt.max_load())
         {  // prevent rehashing during swapping, as it invalidates iterators
